refactor: Move hash functions into utils
