+++ backend/src/core/middlewares/auth.middleware.ts
@@ -4,8 +4,9 @@ import bcrypt from 'bcrypt'
 import { User } from '@core/models'
 import { HashedOtp, VerifyOtpInput } from '@core/interfaces'
 import logger from '@core/logger'
-import { otpClient, mailClient, hashService } from '@core/services'
+import { otpClient, mailClient } from '@core/services'
 import config from '@core/config'
+import { randomSalt } from '@core/utils/hash'
 
 const { retries: OTP_RETRIES, expiry: OTP_EXPIRY, resendTimeout: OTP_RESEND_TIMEOUT } = config.otp
 
@@ -121,7 +122,7 @@ const getOtp = async (req: Request, res: Response): Promise<Response> => {
   }
   try {
     const otp = generateOtp()
-    const hashValue = await hashService.randomSalt(otp)
+    const hashValue = await randomSalt(otp)
     const hashedOtp: HashedOtp = { hash: hashValue, retries: OTP_RETRIES, createdAt: Date.now() }
     await saveHashedOtp(email, hashedOtp)
     await sendOtp(email, otp)
