+++ backend/src/core/services/index.ts
@@ -4,4 +4,3 @@ export * from './campaign.service'
 export * from './job.service'
 export * from './template.service'
 export * from './credential.service'
-export * from './hash.service'
